refactor(mcp): defineMCPTool helper reduces per-tool boilerplate
Move each MCP tool's name/description/inputSchema next to its handler via a
new runtime helper. toolRegistry.ts becomes a thin import-and-list manifest
(89 LoC vs the previous 903); per-tool entry shrinks from a ~23 LoC
{ definition, handler } literal with an unsafe-cast wrapper to a single
identifier in the registry array.

- src/agent/mcp/defineMCPTool.ts (new, 80 LoC): typed wrapper that produces
  ToolRegistryEntry from { name, description, inputSchema, handler<I> } plus
  optional ToolMetadata (requiresAuth, mutatesSession, category,
  defaultErrorRecovery) for agent-facing introspection.
- Each tools/<name>.ts now exports a <toolName>McpTool descriptor alongside
  its handler. 37 tools migrated; no bespoke tool was left out.
- McpToolDefinition / ToolHandler types stay exported from toolRegistry.ts
  (public contract — kernelCAD-server consumes them via the submodule).
- Centralises the `input as Parameters<typeof X>[0]` cast in one place
  instead of 37 sites.

Public API unchanged: TOOL_REGISTRY entry shape, TOOLS, callMcpTool,
getToolDefinition all preserve their existing signatures. All MCP tests pass
(247 / 247 MCP-targeted tests); the 5 spawn.test.ts failures observed in
this worktree are pre-existing dist/cli/replicad_single.wasm-missing
environment issues, not caused by this change.
