+++ src/agent/mcp/tools/addFeature.ts
@@ -2,6 +2,7 @@
 import { addFeature } from '../edits/addFeature';
 import { evaluateScriptTool } from './evaluateScript';
 import type { CompilerDiagnostic } from '../../../shared/diagnostics/diagnostic';
+import { defineMCPTool } from '../defineMCPTool';
 
 export interface AddFeatureInput {
   code: string;
@@ -27,3 +28,18 @@ export async function addFeatureTool(input: AddFeatureInput): Promise<AddFeature
     diagnostics: evalResult.diagnostics,
   };
 }
+
+export const addFeatureMcpTool = defineMCPTool<AddFeatureInput>({
+  name: 'add_feature',
+  description:
+    'Insert a new feature line into a kernelCAD script before the last top-level return statement. Returns the modified code as text plus diagnostics from re-evaluating the result. Side-effect-free. Primitives that accept faceLabels (box, cylinder, extrudeRect, extrudeCircle, extrudePolygon, extrudeRoundedRect) can receive `opts.faceLabels` in the inserted code — use `list_api` to see `featureKindFaceLabels` for the full value schema.',
+  inputSchema: {
+    type: 'object',
+    properties: {
+      code: { type: 'string', description: 'The .kcad.ts source code.' },
+      feature_code: { type: 'string', description: 'Single-statement source line to insert (e.g. `const hole = cylinder(5, 2).translate(10, 10, -1);`).' },
+    },
+    required: ['code', 'feature_code'],
+  },
+  handler: addFeatureTool,
+});
