+++ src/agent/mcp/tools/removeFeature.ts
@@ -2,6 +2,7 @@
 import { removeFeature } from '../edits/removeFeature';
 import { evaluateScriptTool } from './evaluateScript';
 import type { CompilerDiagnostic } from '../../../shared/diagnostics/diagnostic';
+import { defineMCPTool } from '../defineMCPTool';
 
 export interface RemoveFeatureInput {
   code: string;
@@ -29,3 +30,18 @@ export async function removeFeatureTool(
     diagnostics: evalResult.diagnostics,
   };
 }
+
+export const removeFeatureMcpTool = defineMCPTool<RemoveFeatureInput>({
+  name: 'remove_feature',
+  description:
+    'Remove a single line from a kernelCAD script identified by a substring match. Returns the modified code plus diagnostics from re-evaluating. Refuses to remove the line containing the return statement. Side-effect-free.',
+  inputSchema: {
+    type: 'object',
+    properties: {
+      code: { type: 'string', description: 'The .kcad.ts source code.' },
+      match: { type: 'string', description: 'A substring that uniquely identifies the line to remove (e.g. `const hole = cylinder(5,`).' },
+    },
+    required: ['code', 'match'],
+  },
+  handler: removeFeatureTool,
+});
